# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Contacts API
  description: Simple Contacts API with Swagger documentation
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /contacts:
    get:
      summary: Get all contacts
      responses:
        "200":
          description: A list of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      summary: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      summary: Get a contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Contact not found
    put:
      summary: Update a contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Contact updated successfully
        "404":
          description: Contact not found
    delete:
      summary: Delete a contact by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact deleted successfully
        "404":
          description: Contact not found
components:
  schemas:
    Contact:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        favoriteColor:
          type: string
        birthday:
          type: string
          format: date
    ContactInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        favoriteColor:
          type: string
        birthday:
          type: string
          format: date
      required:
        - firstName
        - lastName
        - email